add /api/v1/login, add JWT creation upon login, update schemas to use i64 for all utc fields and nonces, updated migration

+++ crates/db/src/models/badges.rs
@@ -8,5 +8,5 @@ pub struct Badges{
     badge_id: i32,
     badge_description: String,
     badge_url: String,
-    created_utc: i32,
+    created_utc: i64,
 }
