+++ crates/db/src/models/contributors.rs
@@ -6,7 +6,7 @@ pub struct Contributors {
     id: i32,
     user_id: i32,
     board_id: i32,
-    created_utc: i32,
+    created_utc: i64,
     is_active: bool,
     approving_mod_id: i32,
 }
